refactor: TIO-226: extract active reference sound button
